implementing functionality described in issue #20

+++ python3_package_for_geonode/geonode_conabio/import_raster.py
@@ -75,17 +75,22 @@ def main():
     output_filename_temporal = os.path.join(input_directory, filename)
     
     output_filename_temporal += ".tif"
-    
-    with rasterio.open(input_filename) as src:   
-        reproj_and_write_one_band_raster(src, output_filename_temporal)
-    
-    
-    result_import = import_layers_via_docker(region, name_geonode, title,
-                                             abstract, kw,
-                                             output_filename_temporal
-                                             )
-    
-    print(result_import)
-    
-    os.remove(output_filename_temporal)
+    message = """Raster has pixel values that are greater than 255 value, 
+                 then either relabel it to have values under 255 and it can be written 
+                 with a dtype of uint8 -uses less memory- or use different command line
+              """    
+    with rasterio.open(input_filename) as src:
+        array = src.read()
+        unique_values = np.unique(array)
+        del array
+        if np.any(unique_values > 255):
+            print(message)
+        else:
+            reproj_and_write_one_band_raster(src, output_filename_temporal)
+            result_import = import_layers_via_docker(region, name_geonode, title,
+                                                     abstract, kw,
+                                                     output_filename_temporal
+                                                     )
+            print(result_import)
+            os.remove(output_filename_temporal)
